+++ R/reference.R
@@ -6,24 +6,26 @@
 #' @details
 #'  The fields in the `reference` table are:
 #'  \describe{
-#'   \item{\code{rid}}{reference id, a numeric vector}
-#'   \item{\code{type}}{reference type, a factor with levels \code{} \code{article} \code{book} \code{incollection}}
-#'   \item{\code{author}}{author, a character vector}
-#'   \item{\code{title}}{title of publication, a character vector}
-#'   \item{\code{journal}}{journal name, for an \code{article}, character vector}
-#'   \item{\code{month}}{month name, a character vector}
-#'   \item{\code{year}}{a numeric vector}
-#'   \item{\code{volume}}{a character vector}
-#'   \item{\code{number}}{a character vector}
-#'   \item{\code{pages}}{a character vector}
-#'   \item{\code{publisher}}{a character vector}
-#'   \item{\code{address}}{a character vector}
-#'   \item{\code{editor}}{a character vector}
-#'   \item{\code{booktitle}}{a character vector}
-#'   \item{\code{bibtexkey}}{BibTeX key used in the source .bib files, a character vector}
-#'   \item{\code{abstract}}{a character vector}
-#'   \item{\code{note}}{a character vector}
+#'   \item{`rid`}{reference id, a numeric vector}
+#'   \item{`type`}{reference type, a factor with levels \code{} `article` `book` `incollection`}
+#'   \item{`author`}{author, a character vector}
+#'   \item{`title`}{title of publication, a character vector}
+#'   \item{`journal`}{journal name, for an `article`, character vector}
+#'   \item{`month`}{month name, a character vector}
+#'   \item{`year`}{a numeric vector}
+#'   \item{`volume`}{a character vector}
+#'   \item{`number`}{a character vector}
+#'   \item{`pages`}{a character vector}
+#'   \item{`publisher`}{a character vector}
+#'   \item{`address`}{a character vector}
+#'   \item{`editor`}{a character vector}
+#'   \item{`booktitle`}{a character vector}
+#'   \item{`bibtexkey`}{BibTeX key used in the source .bib files, a character vector}
+#'   \item{`abstract`}{a character vector}
+#'   \item{`note`}{a character vector}
 #'  }
+#'
+#'  In the `note` field, `Loc{}` refers to a library catalog identifier, e.g, `Loc{BL: IA 3Q024}`
 #'
 #'  TODO: need a function to turn these back to BibTeX
 #'
@@ -60,5 +62,5 @@ reference <- function() {
 #'   \item{abstract}{abstract, a character vector}
 #'   \item{note}{notes, a character vector}
 #' }
-#' @source \url{https://datavis.ca/milestones/}
+#' @source <https://datavis.ca/milestones/>
 "reference"