+++ NEWS.md
@@ -8,6 +8,10 @@
 * Implement `print_milestone()` method with helper functions
 * Added search_* functions
 * Display media items [link|image] in the print method 
+* Added `print_milestones_cli()` to use the {cli} package for formatting text to console.
+* Fixed NULL values in references and milestones tables: NULL -> NA
+* Converted ropxygen documentation to use markdown, via `roxygen2md::roxygen2md()`
+
 
 ## milestoneR 0.0.0.9000